+++ src/DonorProfiles/Tabs/ProfileTab/LocationRoute.php
@@ -33,7 +33,7 @@ class LocationRoute extends RouteAbstract {
 	 *
 	 * @since 2.10.0
 	 */
-	public function handleRequest( WP_REST_Request $request ) {
+	public function handleRequest( $request ) {
 		return [
 			'states' => LocationList::getStates(
 				$request->get_param( 'countryCode' )