+++ src/DonorProfiles/Tabs/ProfileTab/ProfileRoute.php
@@ -45,7 +45,7 @@ class ProfileRoute extends RouteAbstract {
 	 *
 	 * @since 2.11.0
 	 */
-	public function handleRequest( WP_REST_Request $request ) {
+	public function handleRequest( $request ) {
 		return $this->updateProfile( $request->get_param( 'data' ), $request->get_param( 'id' ) );
 	}